scope: add resolver to handle missing symbols
If we fail to locate a symbol, allow a resolver to generate the symbol.

If successful, it will cache the symbol result in the top-most scope.
We might want to cache the symbol in the scope that generated it, but
it is not yet clear to me if that is the right idea. So I'll be a bit
conservative and stash it in the scope closest to access.