eval: allow some basic calling of raw GI functions
If we have a GIFunction that is part of a namespace, we should
be able to call that function. One such example would be
GObject.type_is_a(GType,GType). This makes that particular
function work.

This GI call dispatch function is getting pretty hairy, but that
is a problem for another day.