tests: add simple test for script
There is a bunch I'd like to do to make this better now, but we have to
get some basics in place and round off some corners which never really
worked with GITypelib access.

My goal here is to make this a bit easier to use as a basic scripting
layer for Builder to use in it's shortcut bindings.