expr: implement user function definitions
We always had this in the syntax, but we never actually finished the
implementation of defining functions so that the user can call them.

I'd still like to change the syntax somewhat, but this at least gets
things working so we can define those functions and use them from the
test script.